handle adoption during deprovisioning
This change updates the ironic provisioner to handle adopting a host
by recreating it in ironic when the ironic database is wiped while the
host is being deprovisioned.

The changes are limited to the ironic package because the more general
adoption logic that works for hosts in steady states does not work for
an active state like deprovisioning.

The first aspect of the change is to ensure that when a host is
registered with ironic, we do as much as possible to ensure it has an
image source. In particular, if there is no Spec.Image because the
host is being deprovisioned, we use the Status.Provisioning.Image
values if they are set. This allows us to register the host with
ironic and then trigger deprovisioning, which is not possible if there
is no image source set.

The second aspect of the change is the handling of the ironic node
states in Deprovision(). The switch statement there is updated to
include the states a node goes through as it is being adopted,
including the fact that it will end up in the "manageable" state for
two different reasons that must be handled separately.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>